requirements.txt
# Smart Cargo Care - Python Dependencies

# Core Computer Vision & Object Detection
ultralytics>=8.0.0          # YOLOv8 object detection framework
opencv-python>=4.8.0        # OpenCV for video processing
numpy>=1.24.0               # Numerical operations

# Serial Communication
pyserial>=3.5               # UART communication with MCU

# Configuration & Data Processing
pyyaml>=6.0                 # YAML configuration file parsing
dataclasses>=0.6            # Data classes for Python 3.7+

# Hugging Face Integration
huggingface_hub             # Hugging Face model download & caching
scipy>=1.10.0               # Hungarian Algorithm for tracking


# Optional: Performance & Visualization
# torch>=2.0.0              # PyTorch (automatically installed with ultralytics)
# matplotlib>=3.7.0         # Plotting and visualization
# pillow>=10.0.0            # Image processing utilities

# Optional: Testing
# pytest>=7.4.0             # Testing framework
# pytest-cov>=4.1.0         # Code coverage
